Prototype build and install script for OpenSubdiv
This script is intended to make building OpenSubdiv with all its various options easier to manage.  It will automatically download and install the dependencies where it can (e.g., it can get GLEW automatically, but it can't get the CUDA SDK).

This is a fork of the similar script found in the USD repo.