Allow using docutils on Windows
On Windows we now look for and use an explicit python interpreter to
find the version of docutils installed.  There is more information in
the commit, but it would result in a much cleaner change if we could
decide that it's okay to do this for all platforms.  For now, just do it
for Windows to get us unblocked.