Further simplification of OpenGL import targets
The internal GL API Loader implementation is self contained and does
not require the system OpenGL implementation at build time, so we
don't need the OpenGL::GL import target for this case.

An earlier change added the OpenGL::GL import target for legacy uses
of GLEW and it remains correct there.

Fixes #1277